# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Seal Online Wiki API
  version: 0.1.0
  summary: Read-only API contract for the first backend iteration.
  description: >
    Initial API design for serving imported Seal Online monster data from the
    normalized database. This first version is intentionally small and focuses
    on read-only endpoints for monsters and maps.

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Monsters
    description: Monster discovery and detail endpoints
  - name: Maps
    description: Map listing endpoints

paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Health check
      description: Simple endpoint to verify the API is running.
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/monsters:
    get:
      tags:
        - Monsters
      summary: List monsters
      description: >
        Returns a paginated list of monsters. Supports lightweight filtering for
        the first backend iteration.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          description: Case-insensitive name search
          schema:
            type: string
        - name: element
          in: query
          description: Filter by element or attribute name
          schema:
            type: string
        - name: level_min
          in: query
          description: Minimum monster level
          schema:
            type: integer
            minimum: 1
        - name: level_max
          in: query
          description: Maximum monster level
          schema:
            type: integer
            minimum: 1
        - name: map_slug
          in: query
          description: Filter monsters found in a specific map
          schema:
            type: string
        - name: source_page
          in: query
          description: Filter by imported level page, for example `1-15`
          schema:
            type: string
      responses:
        '200':
          description: Monster list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonsterListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/monsters/{slug}:
    get:
      tags:
        - Monsters
      summary: Get monster detail
      description: Returns one monster with maps and categorized drops.
      parameters:
        - $ref: '#/components/parameters/MonsterSlug'
      responses:
        '200':
          description: Monster detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonsterDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/maps:
    get:
      tags:
        - Maps
      summary: List maps
      description: Returns a paginated list of maps used by monster locations.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          description: Case-insensitive map name search
          schema:
            type: string
      responses:
        '200':
          description: Map list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number, starting from 1
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    MonsterSlug:
      name: slug
      in: path
      required: true
      description: Unique monster slug
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidPage:
              value:
                error:
                  code: bad_request
                  message: page must be greater than or equal to 1
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            monsterNotFound:
              value:
                error:
                  code: not_found
                  message: Monster not found
    InternalServerError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    PaginationMeta:
      type: object
      required:
        - page
        - page_size
        - total_items
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string

    MapSummary:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        image_url:
          type:
            - string
            - 'null'
          format: uri
        source_url:
          type:
            - string
            - 'null'
          format: uri

    DropEntity:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        image_url:
          type:
            - string
            - 'null'
          format: uri
        source_url:
          type:
            - string
            - 'null'
          format: uri

    MonsterSummary:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        level:
          type:
            - integer
            - 'null'
          minimum: 1
        element:
          type:
            - string
            - 'null'
        image_url:
          type:
            - string
            - 'null'
          format: uri
        source_page:
          type:
            - string
            - 'null'
        source_url:
          type:
            - string
            - 'null'
          format: uri

    MonsterDetail:
      allOf:
        - $ref: '#/components/schemas/MonsterSummary'
        - type: object
          required:
            - found_at
            - drops
          properties:
            found_at:
              type: array
              items:
                $ref: '#/components/schemas/MapSummary'
            drops:
              type: object
              required:
                - gems
                - items
                - quest_items
                - weapons
                - armors
                - accessories
                - b_pet_eq
                - pets
              properties:
                gems:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'
                quest_items:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'
                weapons:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'
                armors:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'
                accessories:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'
                b_pet_eq:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'
                pets:
                  type: array
                  items:
                    $ref: '#/components/schemas/DropEntity'

    MonsterListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MonsterSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    MapListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MapSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'